pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gemini-youtube-transcriber"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "google-genai",
    "mcp[cli]",
    "click",
]

[project.scripts]
transcribe = "youtube_transcriber.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/youtube_transcriber"]
